7,8

+++ desafio8.sql
@@ -1,14 +1,9 @@
-USE SpotifyClone;
-
 DELIMITER $$
-CREATE TRIGGER trigger_usuario_delete
-  BEFORE DELETE ON usuario
-  FOR EACH ROW 
+CREATE TRIGGER  trigger_usuario_delete
+BEFORE DELETE ON usuario
+FOR EACH ROW
 BEGIN
-  DELETE FROM historico
-  WHERE user_id = OLD.user_id;
-  DELETE FROM  seguidores
-  WHERE user_id = OLD.user_id;
-END
-$$
+DELETE FROM SpotifyClone.historico WHERE user_id = OLD.user_id;
+DELETE FROM SpotifyClone.seguidores WHERE user_id = OLD.user_id;
+END $$
 DELIMITER ;
